+++ ChangeLog.md
@@ -5,6 +5,10 @@ SPDX-License-Identifier: Apache-2.0
 
 # Tethys.SPDX
 
+## Next
+
+* `Tethys.SPDX.KnownLicenses`: use `System.Text.Json` instead of `Newtonsoft.Json`.
+
 ## 2.0.0 (2024-02-21)
 
 * Incompatible changes: